requirements.txt
asgiref==3.5.2
attrs==21.4.0
backcall==0.2.0
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.6
django-cache-url==3.4.2
django-extensions==3.2.0
django-widgets-improved==1.5.0
djangorestframework==3.13.1
entrypoints==0.4
env==0.1.0
environs==9.5.0
executing==0.8.3
fastjsonschema==2.15.3
ipython-genutils==0.2.0
Jinja2==3.1.2
jsonschema==4.6.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
MarkupSafe==2.1.1
marshmallow==3.17.0
matplotlib-inline==0.1.3
mistune==0.8.4
nest-asyncio==1.5.5
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.2.1
pywin32==304
pywinpty==2.0.5
pyzmq==23.1.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.2.2.post1
tzdata==2022.1
wcwidth==0.2.5
webencodings==0.5.1
